Deletion of items Bugfix.

+++ ERCentralAPI/ERxWebClient2/ClientApp/app/components/main/mainfull.component.ts
@@ -216,14 +216,17 @@ export class MainFullReviewComponent implements OnInit, OnDestroy {
 		return false;
 	}
 	public DeleteRelevantItems() {
+		if (this.ItemListService.SelectedItems != null &&
+			this.ItemListService.SelectedItems.length > 0) {
 		this.AllIncOrExcShow = false;
 		this.ConfirmationDialogService.confirm("Delete the selected items?",
 			"Are you sure you want to delete these " + this.ItemListService.SelectedItems.length  + " item/s?", false, '')
 			.then((confirm: any) => {
-				if (confirm) {
-					this.ItemListService.DeleteSelectedItems(this.ItemListService.SelectedItems);
+					if (confirm) {
+						this.ItemListService.DeleteSelectedItems(this.ItemListService.SelectedItems);
 				}
-			});
+				});
+		}
 	}
 	public AllocateChoice: string = '';
 	public AllIncOrExcShow: boolean = false;
